% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computesd.R
\name{computesd}
\alias{computesd}
\title{\code{computesd}: Computes standard deviation (SD) by variable type for a data frame}
\usage{
computesd(df)
}
\arguments{
\item{df}{a data frame of character(s), factor(s) or numeric vector(s) of values.}
}
\value{
\code{computesd} returns a table of standard deviations for all columns in the dataframe.
}
\description{
\code{computesd} computes standard deviation for nominal, ordinal or ratio/interval scaled variables.
For a data frame with a variety of data types in the columns standard deviation values are returned.
}
\details{
\code{computesd} produces a table of standard deviations for all columns in the dataframe. If a column is a nominal variable with p categories
then (p-1) standard deviations are returned. If the column is ordinal then the standard deviation is
calculated using 'Agresti' method.
}
\examples{
x1 <- c("b","c","c","b","b","a","b","c","b","a")
x2 <- rnorm(10,35,5)
x3 <- as.ordered(c(1,4,1,3,4,2,4,3,3,4))
x4 <- as.factor(c("L","M","L","H","H","L","H","M","L","H"))
mydat <- data.frame(x1,x2,x3,x4)
}
\keyword{internal}
