% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdfun.R
\name{sdfun}
\alias{sdfun}
\title{\code{sdfun}: Computes standard deviation (SD) of a vector of either a nominal or an ordinal or a numeric variable}
\usage{
sdfun(col)
}
\arguments{
\item{col}{a character, factor or numeric vector of values representing data.}
}
\value{
\code{sdfun} returns standard deviations for the data.
}
\description{
\code{sdfun} computes SD for nominal, ordinal or ratio/interval scaled variables.
For any given set of values (dataframe), SD values are returned.
}
\details{
\code{sdfun} produces standard deviations for the data. If the data are nominal then standard deviation of
total variation is returned unlike 'computesd' function. If the data are ordinal then the standard deviation
returned is calculated using 'Agresti' method.
}
\examples{
x1 <- c("b","c","c","a","a","a","b","c","b","a")
x2 <- rnorm(10,35,5)
x3 <- as.ordered(c(1,4,1,3,4,2,4,3,3,4))
}
\keyword{internal}
