% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcolrm-package.R
\docType{package}
\name{stcolrm-package}
\alias{stcolrm-package}
\title{stcolrm: Calculating standardised coefficients for logistic regression models}
\description{
`stcolrm` is an R package designed to compute standardised coefficients for logistic regression models. It provides a straightforward and statistically sound method to assess the relative importance of predictors by rescaling them to a common scale. This allows for more interpretable comparisons between covariates in binary outcome models. The package uses glm()-fitted logistic regression objects, calculates standardised beta coefficients, and help visualising effect sizes.
}
\author{
\strong{Maintainer}: Asaduzzaman Md \email{asad.mdaz@gmail.com}

}
\keyword{internal}
