% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcolrm.R
\name{stcolrm}
\alias{stcolrm}
\title{\code{stcolrm}: Calculates the standardised coefficients after fitting a logistic regression model}
\usage{
stcolrm(
  object,
  method = c("simple-preSTD", "simple-postEST", "partial", "full"),
  data,
  conf_level = 0.95,
  coef_conversion = exp
)
}
\arguments{
\item{object}{an object of class "formula"  (a symbolic description of the model to be fitted) or 'glm' (a pre-trained binomial model)}

\item{method}{the method ("simple-preSTD", "simple-postEST", "partial", "full") to be used for estimating the standardised coefficients after fitting the model.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.}

\item{conf_level}{a numeric value between 0 and 1 giving the confidence level for the intervals (default is 0.95).}

\item{coef_conversion}{a function to convert the coefficients (default is exp for odds ratios).}
}
\value{
A numeric vector of standardised coefficients of the fitted logistic regression model.
}
\description{
\code{stcolrm} is used to calculate the standardised coefficients after fitting a logistic regression model for a wide
variety of explanatory variables.
}
\details{
stcolrm produces standardised coefficients after fitting a logistic regression model for a wide variety of explanatory variables. It also shows the base results for \code{glm} fit, and also shows the odds ratios of the estimated coefficients. The odds ratios for standardised coefficients are calculated using delta method.
}
\examples{
library(kmed)
heartdata <- heart
heartdata$classN <- ifelse(heart$class == 0, 0, 1)
mfit <- stcolrm(classN ~ age + sex + trestbps + chol + restecg, data=heartdata, "simple-preSTD")
mfit

mfit <- stcolrm(classN ~ age + sex + trestbps + chol + restecg, data=heartdata, "simple-postEST")
mfit

mfit <- stcolrm(classN ~ age + sex + trestbps + chol + restecg, data=heartdata, "partial")
mfit

mfit <- stcolrm(classN ~ age + sex + trestbps + chol + restecg, data=heartdata, "full")
mfit

}
