% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcolrm_plot.R
\name{stcolrm_plot}
\alias{stcolrm_plot}
\title{Visualisation of standardised or raw odds ratios from a stcolrm object}
\usage{
stcolrm_plot(object, .type = c("std", "raw"), log_10 = TRUE)
}
\arguments{
\item{object}{A stcolrm object}

\item{.type}{Type of odds ratios to plot, either "std" for standardised or "raw" for raw odds ratios. Default is "std".}

\item{log_10}{Logical, if TRUE (default) the x-axis is on a log10 scale.}
}
\value{
A ggplot2 object visualising the odds ratios with confidence intervals.
}
\description{
Visualisation of standardised or raw odds ratios from a stcolrm object
}
