% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth1.R
\docType{data}
\name{synth1}
\alias{synth1}
\title{Synthetic Data Set 1}
\format{
A data frame with 1000 rows and 5 variables:
#' \describe{
\item{Var1}{A numeric predictor variable}
\item{Var2}{A numeric predictor variable}
\item{Var3}{A numeric predictor variable}
\item{n}{A numeric variable representing sample size for each observation}
\item{y}{A binomial response variable (0/1) of \code{n} trials}
}
}
\usage{
synth1
}
\description{
A synthetic data set to validate stcolrm.  It has been generated
such that \code{Var1} and \code{Var2} have no effect on the binomial response \code{y},
while \code{Var3} has a positive effect.
}
\keyword{datasets}
